The primary change in this commit is to add a forward inference engine for handling owl:subClassOf. This is very much the bare bones for such an engine. We've incorporated no multithreading. It is a simple depth first search of the subClass graph that populates it all the members of the taxonomy. It is a start.
This commit also updates Printer.java because of bug where it wasn't
printing predicates that were subsequent to the first predicate on a
subject. It now prints them.

Related to another change, the statement index will now only add
statements which are not redundant, i.e. have the same subject
predicate and object. This required an implementation of the equals
method for Statements, which is the change made to Statement.java.
Additionallly we have updated addStatement in
StatementIndex.java to handle cases of repeated statements.
That file has also had the main method removed since it was redundant.
This file has also been updated because we needed an easier access
to the graph created by certain predicates, e.g. owl:subClassOf. So
this class now has a new index for such graphs.